Add files via upload
A company offers premium membership to customers based on the following criteria:

The customer must be at least 18 years old.
Their monthly income should be greater than or equal to ₹30,000.
They should not have any pending dues.
Write a program that uses relational, logical, and ternary operators to check if a customer is eligible for premium membership. Use appropriate keywords, identifiers, and access modifiers for the program.